feat!: use Decimal for monetary and quantity fields (#91)
Every monetary and quantity field was parsed to an IEEE double. Coinbase
sends these as decimal strings, so converting them lost precision before
the caller ever saw the value, with no way to opt out.

Product.quoteIncrement and baseIncrement are the tick sizes every order
must be quantized to; as doubles (0.01, 0.00000001) they could not be
used for exact quantization. Balances such as 61250.10 drifted under
arithmetic. Callers needing exactness had to bypass the models and
re-parse the raw JSON.

Money is now Decimal (package:decimal), never double:

- 42 double fields across Product, Order, Fill, Account, Ticker, Trade,
  PreviewOrderResponse and all five order configurations.
- 19 num fields across TransactionSummary, SpotPosition and
  VolumeBreakdown.
- Order.numberOfFills is now int?, matching what it is: a count.
- getAccountBalance returns Future<Decimal?>.
- toJson()/toCBJson() serialize decimals as strings, so payloads stay
  jsonEncode-able and round-trip without loss.

tools.dart gains nullableDecimal, requiredDecimal and nullableInt.
nullableDouble is removed. nullableNumber is kept for ServerTime alone,
whose epochMillis arrives fractional ("1715591792.118") -- a timestamp,
not money.

Note that Decimal normalizes trailing zeros: a wire value of '61250.10'
parses to the exact value 61250.1. The value is preserved exactly; only
its textual form is normalized.

Also fixes two pre-existing null-safety bugs found while building the
round-trip acceptance test, both of which threw
"type 'Null' is not a subtype of type 'String'":

- Account.fromJson parsed createdAt, updatedAt and deletedAt without a
  null guard. deletedAt is null for every account that has not been
  deleted, so round-tripping a live account through toJson()/fromJson()
  always threw. All three are now guarded, matching fromCBJson.
- The eight order enums' fromCB rejected a null value, so
  Order.fromCBJson could not parse an order omitting trigger_status,
  order_type, reject_reason or product_type. Each now accepts String?
  and returns its unknown/unspecified member, matching the existing
  OrderRejectReason.fromCB precedent and the behaviour for an
  unrecognised string.

Adds decimal: ^3.2.6. Bumps version to 0.10.0 and documents the
money-is-never-a-double contract in the README.

BREAKING CHANGE: monetary and quantity fields are Decimal, not double.
Order.numberOfFills is int?. getAccountBalance returns Future<Decimal?>.
nullableDouble is removed; use nullableDecimal, nullableInt or
nullableNumber. Convert at your application's edge with .toDouble().

+++ lib/src/rest/accounts.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:decimal/decimal.dart';
 import 'package:coinbase_cloud_advanced_trade_client/src/models/account.dart';
 import 'package:coinbase_cloud_advanced_trade_client/src/models/credential.dart';
 import 'package:coinbase_cloud_advanced_trade_client/src/models/error.dart';
@@ -171,9 +172,9 @@ Future<Account?> getAccount(
 /// [credential] - The user's API credentials.
 /// [isSandbox] - Whether to use the sandbox environment.
 ///
-/// Returns the available balance of the account as a double, or null if the
-/// account is not found.
-Future<double?> getAccountBalance(
+/// Returns the available balance of the account as a [Decimal], or null if
+/// the account is not found.
+Future<Decimal?> getAccountBalance(
     {String? uuid,
     String? currency,
     http.Client? client,
